+++ Smssme/Controller/TabBarController.swift
@@ -18,17 +18,7 @@ class TabBarController: UITabBarController {
         showFirstView()
     }
     
-    // 로그인 유무에 따라 앱 실행 시 처음 보여줄 탭 설정
-    private func showFirstView() {
-        if let user  = Auth.auth().currentUser {
-            self.selectedViewController = viewControllers?[0]
-        } else {
-            self.selectedViewController = viewControllers?[2]
-        }
-    }
-    
     func configureController() {
-        //메인페이지
         let mainPage = tabBarNavigationController(
             unselectedImage: UIImage(systemName: "house.fill") ?? UIImage(),
             selectedImage: UIImage(systemName: "house.fill") ?? UIImage(),
@@ -36,7 +26,6 @@ class TabBarController: UITabBarController {
             rootViewController: MainPageVC()
         )
         
-        //머니다이어리
         let diary = tabBarNavigationController(
             unselectedImage: UIImage(systemName: "calendar") ?? UIImage(),
             selectedImage: UIImage(systemName: "calendar") ?? UIImage(),
@@ -48,20 +37,9 @@ class TabBarController: UITabBarController {
             unselectedImage: UIImage(systemName: "note.text.badge.plus") ?? UIImage(),
             selectedImage: UIImage(systemName: "note.text.badge.plus") ?? UIImage(),
             isNavigationBarHidden: false,
-            rootViewController: {
-                let planService = FinancialPlanService()
-                let plans = planService.fetchAllFinancialPlans()
-                
-                if plans.isEmpty {
-                    return FinancialPlanSelectionVC()
-                } else {
-                    let firstPlan = plans.first!
-                    return FinancialPlanCurrentPlanVC(planService: planService, planDTO: firstPlan)
-                }
-            }()
+            rootViewController: planPageCondition()
         )
         
-        // 마이페이지
         let myPage = tabBarNavigationController(
             unselectedImage: UIImage(systemName: "person.and.background.striped.horizontal") ?? UIImage(),
             selectedImage: UIImage(systemName: "person.and.background.striped.horizontal") ?? UIImage(),
@@ -82,3 +60,27 @@ class TabBarController: UITabBarController {
     }
 }
 
+// MARK: - 페이지 분기 처리
+extension TabBarController {
+    // 로그인 유무에 따라 앱 실행 시 처음 보여줄 탭 설정
+    private func showFirstView() {
+        if let user  = Auth.auth().currentUser {
+            self.selectedViewController = viewControllers?[0]
+        } else {
+            self.selectedViewController = viewControllers?[2]
+        }
+    }
+    
+    // 진행중 플랜이 있다면 막대그래프 페이지, 아니면 선택창
+    private func planPageCondition() -> UIViewController {
+        let planService = FinancialPlanService()
+        let plans = planService.fetchIncompletedPlans()
+        
+        if plans.isEmpty {
+            return FinancialPlanSelectionVC()
+        } else {
+            let firstPlan = plans.first!
+            return FinancialPlanCurrentPlanVC(planService: planService, planDTO: firstPlan)
+        }
+    }
+}
